// src/app/model/routing-enums/app-routes.enum.enum.ts
export enum AppRoutes {
  Default = '',
  // Logger = 'admin/logger',
  Welcome = 'welcome',
  Login = 'login',
  Products = 'products',
  User = 'user',
  Category = 'category',
  Merchant = 'merchant',
  Subscriptions = 'subscriptions',
  Customer = 'customer',
  Order = 'order',
  Cart = 'cart',
  Coupon = 'coupon',
  Employee = 'employee',
  Banner = 'banner',
  NotFound = 'not-found',
  Servicableareas = 'servicableareas',
  Invoice = 'invoice',
  Delivery = 'delivery',
  Feedbacks = 'feedbacks',
  // Dashboard = 'user/dashboard'
  // Version = 'version'
}